+++ happyhouse_backend/src/main/java/com/ssafy/happyhouse/model/Bookmark.java
@@ -8,6 +8,6 @@ import lombok.Data;
 public class Bookmark {
 
 	private String userid;
-	private int aptCode;
+	private long aptCode;
 
 }
